pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csp-workflow-engine"
version = "0.1.0"
description = "Graph-based regulatory workflow engine for clinical trial statistical programming"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
csp-validate-graph = "scripts.evaluation:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
csp_engine = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
